search-agent-using-react-architecture/pyproject.toml
[project]
name = "search-agent-using-react-architecture"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=26.1.0",
    "isort>=8.0.1",
    "langchain>=1.2.10",
    "langchain-openai>=1.1.10",
    "langchain-tavily>=0.2.17",
    "python-dotenv>=1.2.1",
    "tavily>=1.1.0",
]
